pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autotext2sql"
version = "0.1.0"
description = "Agentic Text-to-SQL system with metadata search"
requires-python = ">=3.11"
dependencies = [
    # Core framework
    "langgraph==1.1.6",
    "langgraph-checkpoint-sqlite==3.0.3",
    "langchain-core==1.2.29",
    "langchain-openai==1.1.13",
    # API
    "fastapi==0.135.3",
    "uvicorn[standard]==0.44.0",
    "python-multipart==0.0.26",
    # Config
    "pydantic==2.13.0",
    "pydantic-settings==2.13.1",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.2",
    # DB
    "sqlalchemy==2.0.49",
    "psycopg2-binary==2.9.11",
    "pymysql==1.1.2",
    # Vector store
    "qdrant-client==1.17.1",
    # Embeddings
    # SQL parsing / validation
    "sqlglot==30.4.3",
    # Observability
    "structlog==25.5.0",
    "opentelemetry-api==1.41.0",
    "opentelemetry-sdk==1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.41.0",
    "prometheus-client==0.25.0",
    # Token counting
    "tiktoken==0.12.0",
    # HTTP client
    "httpx==0.28.1",
    # LangSmith (optional tracing)
    "langsmith==0.7.31",
    "pyodbc==5.3.0",
    "mem0ai>=1.0.11",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "httpx==0.28.1",
]
ui = [
    "gradio==6.12.0",
]

[project.scripts]
autotext2sql-index = "autotext2sql.index:main"
autotext2sql-serve = "autotext2sql.api:serve"

[tool.hatch.build.targets.wheel]
packages = ["autotext2sql", "evals"]
